added the ConsoleAppController project which has the goal of simply instanciating the services and providing them with the data for my personal project.
in using the program with actual data and runing through loops thousands of times errors have shown themselves and have been solved. thse include:

OrbitItem: having int values for Mass and OrbitingDistance has revealed itself to be way too small. Mass has been converted to a double and Orbiting Distance to a long.

minor correctsions in Tide calculator, the more relavent being that ProvideRangeToWorkWith would pass the actual values along which could course problems if they were edited externally. thus now instead copies are passed along. and safeguards have been added to avoid an angle value that is NaN.

IOrbitItemsRepository has gained the function GetAvailableId(int) which returns a range of available ids which was required in the copying process.

minor updates in tests with some of the changes. added tests in OrbitItemRepositoryLocal/GatAvailableId for ranges.
added tests in TideCalculator/ProvideRangeToWorkWith to test for the new call.